+++ CHANGELOG.md
@@ -8,3 +8,11 @@ v0.2.0 (2018-12-12)
 -	[Preview] Can generate a playblast directly, bypassing the UI.
 -	[Preview] Added preset for proxy resolution.
 
+v0.2.1 (--)
+-	[Preview] Renamed 'Slate' to 'Burn-in'.
+-	[Preview] Revised burn-in HUD info for consistency with dailies burn-in.
+-	[Preview] Enabled dailies publishing.
+-	[Preview] Daily publish option disabled when format is set to QuickTime movie.
+-	[Preview] Added ability to output TIFF images (with alpha channel).
+
+(TODO)